Migrate cookCompilationDiagnostics to sketch
Diagnostics and hover are now dependent on sketch instead of parser for the errorNodes
